refactor: instrumentor agent enabled config calculation
