+++ instrumentor/controllers/agentenabled/dynamicconfig/config.go
@@ -1,0 +1,147 @@
+package dynamicconfig
+
+import (
+	"github.com/odigos-io/odigos/api/k8sconsts"
+	odigosv1 "github.com/odigos-io/odigos/api/odigos/v1alpha1"
+	"github.com/odigos-io/odigos/common"
+	commonapi "github.com/odigos-io/odigos/common/api"
+	commonapisampling "github.com/odigos-io/odigos/common/api/sampling"
+	"github.com/odigos-io/odigos/distros/distro"
+	"github.com/odigos-io/odigos/instrumentor/controllers/agentenabled/dynamicconfig/metrics"
+	"github.com/odigos-io/odigos/instrumentor/controllers/agentenabled/dynamicconfig/traces"
+	"github.com/odigos-io/odigos/instrumentor/controllers/agentenabled/signals"
+	"github.com/odigos-io/odigos/k8sutils/pkg/workload"
+)
+
+// return type from the function that computes the dynamic container configs.
+// dynamic config contains agent signals configs, and collector config (if any).
+type DynamicContainerConfigs struct {
+	// Agent
+	AgentTracesConfig  *odigosv1.AgentTracesConfig
+	AgentMetricsConfig *odigosv1.AgentMetricsConfig
+	AgentLogsConfig    *odigosv1.AgentLogsConfig
+
+	// Collector
+	CollectorConfig *commonapi.ContainerCollectorConfig
+}
+
+// Calculate the dynamic container configs for a given container.
+// the dynamic config contains two parts:
+// - agent (per signal config that is consumed by the selected instrumentation agent)
+// - collector (config for this container used by collector components)
+//
+// this function computes the dynamic configs and returns them in a single struct.
+// if there is any conflict or issue, and the container should be disabled, return the disabled info.
+func CalculateDynamicContainerConfig(
+	containerName string,
+	irls *[]odigosv1.InstrumentationRule,
+	effectiveConfig *common.OdigosConfiguration,
+	runtimeDetails *odigosv1.RuntimeDetailsByContainer,
+	agentLevelActions *[]odigosv1.Action,
+	samplingRules *[]odigosv1.Sampling,
+	workloadObj workload.Workload,
+	pw k8sconsts.PodWorkload,
+	d *distro.OtelDistro,
+	enabledSignals signals.EnabledSignals,
+) (*DynamicContainerConfigs, *odigosv1.AgentDisabledInfo) {
+
+	agentSpanMetricsEnabled := metrics.AgentSpanMetricsEnabled(effectiveConfig)
+
+	var collectorConfig *commonapi.ContainerCollectorConfig
+
+	var tracesConfig *odigosv1.AgentTracesConfig
+	if enabledSignals.TracesEnabled {
+		tracesConfig = &odigosv1.AgentTracesConfig{}
+
+		// Id Generator
+		idGeneratorConfig, err := traces.CalculateIdGeneratorConfig(effectiveConfig)
+		if err != nil {
+			return nil, err
+		}
+		tracesConfig.IdGenerator = idGeneratorConfig // can be nil
+
+		// Url Templatization
+		urlTemplatizationConfig := traces.CalculateUrlTemplatizationConfig(agentLevelActions, containerName, runtimeDetails.Language, pw)
+		if urlTemplatizationConfig != nil {
+			if traces.DistroSupportsTracesUrlTemplatization(d) && agentSpanMetricsEnabled {
+				tracesConfig.UrlTemplatization = urlTemplatizationConfig
+			} else {
+				collectorConfig = &commonapi.ContainerCollectorConfig{
+					UrlTemplatization: urlTemplatizationConfig,
+				}
+			}
+		}
+
+		// Sampling
+		noisyOps, relevantOps, costRules := traces.CalculateSamplingCategoryRulesForContainer(samplingRules, runtimeDetails.Language, pw, containerName, d, workloadObj, effectiveConfig)
+		// if we have any noisy operations, we need to add them to the traces config.
+		// use head/tail sampling based on the distro support.
+		if len(noisyOps) > 0 {
+			if traces.DistroSupportsHeadSampling(d) {
+				tracesConfig.HeadSampling = &odigosv1.HeadSamplingConfig{
+					NoisyOperations: noisyOps,
+				}
+			} else {
+				if collectorConfig == nil {
+					collectorConfig = &commonapi.ContainerCollectorConfig{}
+				}
+				collectorConfig.TailSampling = &commonapisampling.TailSamplingSourceConfig{
+					NoisyOperations: noisyOps,
+				}
+			}
+		}
+		// if we have any highly-relevant or cost-reduction rules, we need to add them to the collector config.
+		// create tail sampling for this source if not already created.
+		if len(relevantOps) > 0 || len(costRules) > 0 {
+			if collectorConfig == nil {
+				collectorConfig = &commonapi.ContainerCollectorConfig{}
+			}
+			if collectorConfig.TailSampling == nil {
+				collectorConfig.TailSampling = &commonapisampling.TailSamplingSourceConfig{}
+			}
+			collectorConfig.TailSampling.HighlyRelevantOperations = relevantOps
+			collectorConfig.TailSampling.CostReductionRules = costRules
+		}
+
+		// Headers Collection - Agent only (not applicable to collector)
+		if traces.DistroSupportsTracesHeadersCollection(d) {
+			tracesConfig.HeadersCollection = traces.CalculateHeaderCollectionConfig(d, irls)
+		}
+
+		// Span Renamer
+		// TODO: add support to do it in the collector
+		if traces.DistroSupportsTracesSpanRenamer(d) {
+			tracesConfig.SpanRenamer = traces.CalculateSpanRenamerConfig(agentLevelActions, runtimeDetails.Language)
+		}
+
+		// Payload Collection - Agent only (not applicable to collector)
+		if traces.DistroSupportsTracesPayloadCollection(d) {
+			tracesConfig.PayloadCollection = traces.CalculatePayloadCollectionConfig(d, irls)
+		}
+	}
+
+	var metricsConfig *odigosv1.AgentMetricsConfig
+	if enabledSignals.MetricsEnabled {
+		metricsConfig = &odigosv1.AgentMetricsConfig{}
+
+		if metrics.DistroSupportsAgentSpanMetrics(d) && agentSpanMetricsEnabled {
+			agentSpanMetricsConfig, err := metrics.CalculateAgentSpanMetricsConfig(effectiveConfig, d)
+			if err != nil {
+				return nil, err
+			}
+			metricsConfig.SpanMetrics = agentSpanMetricsConfig
+		}
+	}
+
+	var logsConfig *odigosv1.AgentLogsConfig
+	if enabledSignals.LogsEnabled {
+		logsConfig = &odigosv1.AgentLogsConfig{}
+	}
+
+	return &DynamicContainerConfigs{
+		AgentTracesConfig:  tracesConfig,
+		AgentMetricsConfig: metricsConfig,
+		AgentLogsConfig:    logsConfig,
+		CollectorConfig:    collectorConfig,
+	}, nil
+}
