--- instrumentor/controllers/agentenabled/sampling/tailsampling.go
@@ -1,57 +1,0 @@
-package sampling
-
-import (
-	"github.com/odigos-io/odigos/api/k8sconsts"
-	odigosv1 "github.com/odigos-io/odigos/api/odigos/v1alpha1"
-	"github.com/odigos-io/odigos/common"
-	apisampling "github.com/odigos-io/odigos/common/api/sampling"
-	"github.com/odigos-io/odigos/distros/distro"
-	"github.com/odigos-io/odigos/k8sutils/pkg/workload"
-)
-
-func FilterTailSamplingRulesForContainer(samplingRules *[]odigosv1.Sampling, language common.ProgrammingLanguage,
-	pw k8sconsts.PodWorkload, containerName string, distro *distro.OtelDistro, workloadObj workload.Workload, effectiveConfig *common.OdigosConfiguration) ([]apisampling.NoisyOperation, []apisampling.HighlyRelevantOperation, []apisampling.CostReductionRule) {
-
-	var filteredNoisyOps []apisampling.NoisyOperation
-	var filteredRelevantOps []apisampling.HighlyRelevantOperation
-	var filteredCostRules []apisampling.CostReductionRule
-
-	for _, samplingRule := range *samplingRules {
-		// Filter and convert NoisyOperations, HighlyRelevantOperations, CostReductionRules.
-		// Exclude SourceScopes and Notes from the rules because we want the instrumentationConfig to be more lightweight.
-
-		// If the distro not supports head sampling, we need the NoisyOperations to be applied at the collector level (tailsampling).
-		if distro.Traces == nil || distro.Traces.HeadSampling == nil || !distro.Traces.HeadSampling.Supported {
-			filteredNoisyOps = CalculateNoisyOperationsForContainer(distro, workloadObj, containerName, effectiveConfig, samplingRules, pw)
-		}
-
-		// Filter and convert HighlyRelevantOperations - exclude SourceScopes and Notes
-		for _, relevantOp := range samplingRule.Spec.HighlyRelevantOperations {
-			if IsServiceInRuleScope(relevantOp.SourceScopes, pw, containerName, language) {
-				filteredRelevantOps = append(filteredRelevantOps, apisampling.HighlyRelevantOperation{
-					Id:                odigosv1.ComputeHighlyRelevantOperationHash(&relevantOp),
-					Name:              relevantOp.Name,
-					Disabled:          relevantOp.Disabled,
-					Error:             relevantOp.Error,
-					DurationAtLeastMs: relevantOp.DurationAtLeastMs,
-					Operation:         relevantOp.Operation,
-					PercentageAtLeast: relevantOp.PercentageAtLeast,
-				})
-			}
-		}
-
-		for _, costRule := range samplingRule.Spec.CostReductionRules {
-			if IsServiceInRuleScope(costRule.SourceScopes, pw, containerName, language) {
-				filteredCostRules = append(filteredCostRules, apisampling.CostReductionRule{
-					Id:               odigosv1.ComputeCostReductionRuleHash(&costRule),
-					Name:             costRule.Name,
-					Disabled:         costRule.Disabled,
-					Operation:        costRule.Operation,
-					PercentageAtMost: costRule.PercentageAtMost,
-				})
-			}
-		}
-	}
-
-	return filteredNoisyOps, filteredRelevantOps, filteredCostRules
-}
